Remove more asserts, further tweaking...

+++ mednafen/sound/Blip_Buffer.cpp
@@ -114,7 +114,7 @@ Blip_Buffer::blargg_err_t Blip_Buffer::set_sample_rate( long new_rate, int msec 
 
 blip_resampled_time_t Blip_Buffer::clock_rate_factor( long rate ) const
 {
-	double ratio = (double) sample_rate_ / rate;
+	float ratio = (float) sample_rate_ / rate;
 	blip_s64 factor = (blip_s64) floor( ratio * (1LL << BLIP_BUFFER_ACCURACY) + 0.5 );
 	return (blip_resampled_time_t) factor;
 }
@@ -185,7 +185,7 @@ Blip_Synth_Fast_::Blip_Synth_Fast_()
 	delta_factor = 0;
 }
 
-void Blip_Synth_Fast_::volume_unit( double new_unit )
+void Blip_Synth_Fast_::volume_unit( float new_unit )
 {
 	delta_factor = int (new_unit * (1L << blip_sample_bits) + 0.5);
 }
